googlecloud/smart-invest/cloud-function/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-logging
functions-framework